mnt/airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests
astronomer-cosmos
astro-sdk-python[postgres]
boto3